Update dummy test to use FutureMatchers and configure circle CI with PG
