+++ .circleci/config.yml
@@ -8,18 +8,22 @@ jobs:
     docker:
     # specify the version you desire here
     - image: circleci/openjdk:8-jdk
+      environment:
+        # PG client params (test conf fetch them from env)
+        PGUSER: signalconso
+        PGHOST: localhost
+        PGDATABASE: test_signalconso
+        # Customize the JVM maximum heap limit
+        JVM_OPTS: -Xmx3200m
+        TERM: dumb
 
-    # Specify service dependencies here if necessary
-    # CircleCI maintains a library of pre-built images
-    # documented at https://circleci.com/docs/2.0/circleci-images/
-    # - image: circleci/postgres:9.4
+    # See https://circleci.com/docs/2.0/postgres-config/
+    - image: circleci/postgres:10.6-ram
+      environment:
+        POSTGRES_USER: signalconso
+        POSTGRES_DB: test_signalconso
 
     working_directory: ~/repo
-
-    environment:
-      # Customize the JVM maximum heap limit
-      JVM_OPTS: -Xmx3200m
-      TERM: dumb
 
     steps:
     - checkout
