+++ conf/test.conf
@@ -5,6 +5,9 @@ test.db {
     host="localhost"
     name="test_signalconso"
 }
+test.db.user = ${?PGUSER}
+test.db.host = ${?PGHOST}
+test.db.name = ${?PGDATABASE}
 
 slick.dbs.default.db.properties.url = "postgres://"${test.db.user}"@"${test.db.host}"/"${test.db.name}
